# contracts/openapi/moneysnap-v1.yaml
openapi: 3.1.0
info:
  title: Money Snap API
  version: 0.1.0
  description: Canonical REST contract for the Money Snap iOS MVP.
servers:
  - url: /api/v1
paths:
...
  /groups:
    post:
      operationId: createGroup
      summary: Create a group owned by the authenticated user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreateRequest'
            examples:
              canonical:
                externalValue: ../examples/v1/groups/create-request.json
      responses:
        '201':
          description: Group created or an identical mutation replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
              examples:
                canonical:
                  externalValue: ../examples/v1/groups/group-response.json
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: The actor already used the mutation ID for a different payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                canonical:
                  externalValue: ../examples/v1/snaps/error-mutation-conflict.json
        '500':
          description: The request could not be safely persisted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                canonical:
                  externalValue: ../examples/v1/snaps/error-internal.json
    get:
      operationId: listGroups
      summary: List groups the authenticated user currently belongs to.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Membership groups.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupListResponse'
              examples:
                canonical:
                  externalValue: ../examples/v1/groups/group-list-response.json
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: The request could not be safely read.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                canonical:
                  externalValue: ../examples/v1/snaps/error-internal.json
  /groups/{groupId}:
    get:
      operationId: getGroup
      summary: Return one group the actor currently belongs to.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Group membership.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
              examples:
                canonical:
                  externalValue: ../examples/v1/groups/group-response.json
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotAccessible'
        '500':
          description: The request could not be safely read.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                canonical:
                  externalValue: ../examples/v1/snaps/error-internal.json
    delete:
      operationId: deleteGroup
      summary: Delete an owned group without deleting personal Snaps.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - name: X-Client-Mutation-Id
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: '.*\S.*'
      responses:
        '204':
          description: Group deleted or an identical delete mutation replayed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotAccessible'
        '409':
          description: The actor already used the mutation ID for a different group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                canonical:
                  externalValue: ../examples/v1/snaps/error-mutation-conflict.json
        '500':
          description: The request could not be safely persisted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                canonical:
                  externalValue: ../examples/v1/snaps/error-internal.json
  /groups/{groupId}/invites:
    post:
      operationId: issueGroupInvite
      summary: Issue or rotate the single active invite for an owned group.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Active invite code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedInviteResponse'
              examples:
                canonical:
                  externalValue: ../examples/v1/groups/invite-issue-response.json
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotAccessible'
    delete:
      operationId: revokeGroupInvite
      summary: Revoke the active invite for an owned group.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '204':
          description: Invite revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotAccessible'
  /groups/{groupId}/members:
    get:
      operationId: listGroupMembers
      summary: List current members of a group the actor belongs to.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Group members.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListResponse'
              examples:
                canonical:
                  externalValue: ../examples/v1/groups/member-list-response.json
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotAccessible'
  /groups/{groupId}/members/me:
    delete:
      operationId: leaveGroup
      summary: Leave a group without deleting personal Snaps.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '204':
          description: Membership removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotAccessible'
  /groups/{groupId}/members/{memberId}:
    delete:
      operationId: removeGroupMember
      summary: Owner removes one member without deleting personal Snaps.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/MemberId'
      responses:
        '204':
          description: Member removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotAccessible'
  /groups/{groupId}/today:
    get:
      operationId: getGroupToday
      summary: Return today's shared projection for a group. Hidden groups omit amount fields.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - name: timeZone
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Visible or hidden group Today projection.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/VisibleGroupTodayResponse'
                  - $ref: '#/components/schemas/HiddenGroupTodayResponse'
              examples:
                visible:
                  externalValue: ../examples/v1/groups/visible-today-response.json
                hidden:
                  externalValue: ../examples/v1/groups/hidden-today-response.json
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotAccessible'
  /groups/{groupId}/members/{memberId}/today:
    get:
      operationId: getGroupMemberToday
      summary: Return one member's today representative in a group.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/MemberId'
        - name: timeZone
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: One member today row.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/VisibleMemberToday'
                  - $ref: '#/components/schemas/HiddenMemberToday'
              examples:
                visible:
                  externalValue: ../examples/v1/groups/member-today-response.json
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotAccessible'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque access token
  responses:
    BadRequest:
      description: Request validation or semantics failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            canonical:
              externalValue: ../examples/v1/identity/error-invalid-request.json
    Unauthorized:
      description: Authentication credentials are missing, invalid, expired, or revoked.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            canonical:
              externalValue: ../examples/v1/identity/error-session-rejected.json
    NotAccessible:
      description: The Snap does not exist or is not owned by the actor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            canonical:
              externalValue: ../examples/v1/snaps/error-not-accessible.json
  parameters:
...
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    GroupCreateRequest:
      type: object
      additionalProperties: false
      required:
        - clientMutationId
        - name
        - amountVisible
      properties:
        clientMutationId:
          type: string
          minLength: 1
          maxLength: 128
          pattern: '.*\S.*'
        name:
          type: string
          minLength: 1
        amountVisible:
          type: boolean
    GroupResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - amountVisible
        - role
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        amountVisible:
          type: boolean
        role:
          type: string
          enum: [owner, member]
        createdAt:
          type: string
          format: date-time
...
    GroupListResponse:
      type: object
      additionalProperties: false
      required:
        - groups
      properties:
        groups:
          type: array
          items:
            $ref: '#/components/schemas/GroupResponse'
    IssuedInviteResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - expiresAt
      properties:
        code:
          type: string
        expiresAt:
          type: string
          format: date-time
...
    MemberListResponse:
      type: object
      additionalProperties: false
      required:
        - members
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/GroupMember'
    GroupMember:
      type: object
      additionalProperties: false
      required:
        - userId
        - displayName
        - avatar
        - role
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
        avatar:
          type: string
        role:
          type: string
          enum: [owner, member]
    VisibleGroupTodayResponse:
      type: object
      additionalProperties: false
      required:
        - localDay
        - members
      properties:
        localDay:
          type: string
          format: date
        members:
          type: array
          items:
            $ref: '#/components/schemas/VisibleMemberToday'
    HiddenGroupTodayResponse:
      type: object
      additionalProperties: false
      required:
        - localDay
        - members
      properties:
        localDay:
          type: string
          format: date
        members:
          type: array
          items:
            $ref: '#/components/schemas/HiddenMemberToday'
    VisibleMemberToday:
      type: object
      additionalProperties: false
      required:
        - userId
        - displayName
        - avatar
        - snapCount
        - totalAmountWon
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
        avatar:
          type: string
        snapCount:
          type: integer
          minimum: 0
        totalAmountWon:
          type: integer
          format: int64
          minimum: 0
        representative:
          $ref: '#/components/schemas/VisibleSnap'
    HiddenMemberToday:
      type: object
      additionalProperties: false
      required:
        - userId
        - displayName
        - avatar
        - snapCount
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
        avatar:
          type: string
        snapCount:
          type: integer
          minimum: 0
        representative:
          $ref: '#/components/schemas/HiddenSnap'
    VisibleSnap:
      type: object
      additionalProperties: false
      required:
        - snapId
        - category
        - amountWon
        - sharedAt
      properties:
        snapId:
          type: string
          format: uuid
        category:
          type: string
          enum: [food, cafe, transportation, shopping, living, culture, health, other]
        amountWon:
          type: integer
          format: int64
        sharedAt:
          type: string
          format: date-time
    HiddenSnap:
      type: object
      additionalProperties: false
      required:
        - snapId
        - category
        - sharedAt
      properties:
        snapId:
          type: string
          format: uuid
        category:
          type: string
          enum: [food, cafe, transportation, shopping, living, culture, health, other]
        sharedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - correlationId
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
          enum:
            - INVALID_REQUEST
            - SESSION_REJECTED
            - MUTATION_CONFLICT
            - SNAP_VERSION_CONFLICT
            - NOT_ACCESSIBLE
            - INVALID_CURSOR
            - INTERNAL_ERROR
            - APPLE_REAUTHENTICATION_REJECTED
            - APPLE_REVOCATION_FAILED
            - APPLE_EVENT_REJECTED
            - APPLE_EVENT_UNSUPPORTED
        correlationId:
          type: string
          format: uuid
          description: Identifier used to correlate the failed request with server logs.